Fix attention precision mismatch align eager_attention_forward with hf
