+++ paddleformers/nn/attention/eager_attention.py
@@ -18,7 +18,14 @@ import paddle
 import paddle.nn as nn
 
 from ...utils.masking_utils import _gen_from_sparse_attn_mask_indices
-from .utils import repeat_kv
+
+
+def repeat_kv(hidden_states: paddle.Tensor, n_rep: int) -> paddle.Tensor:
+    batch, num_key_value_heads, slen, head_dim = hidden_states.shape
+    if n_rep == 1:
+        return hidden_states
+    hidden_states = hidden_states[:, :, None, :, :].expand(batch, num_key_value_heads, n_rep, slen, head_dim)
+    return hidden_states.reshape(batch, num_key_value_heads * n_rep, slen, head_dim)
 
 
 def eager_attention_forward(
@@ -34,9 +41,6 @@ def eager_attention_forward(
     **kwargs,
 ):
 
-    # b h l d -> b l h d
-    key = key.transpose(1, 2)
-    value = value.transpose(1, 2)
     if hasattr(module, "num_key_value_groups"):
         num_key_value_groups = module.num_key_value_groups
         key = repeat_kv(key, num_key_value_groups)
@@ -53,10 +57,7 @@ def eager_attention_forward(
 
         attention_mask = _gen_from_sparse_attn_mask_indices(attn_mask_startend_row_indices, query.dtype, is_causal)
 
-    # b l h d -> b h l d
-    key = key.transpose(1, 2)
-    value = value.transpose(1, 2)
-    attn_weights = paddle.matmul(query, key.transpose([0, 1, 3, 2])) * scaling
+    attn_weights = paddle.matmul(query, key.transpose(2, 3)) * scaling
 
     if attention_mask is not None:
         attention_mask = attention_mask[:, :, :, : key.shape[-2]]
@@ -72,8 +73,8 @@ def eager_attention_forward(
         attn_weights = nn.functional.softmax(attn_weights, axis=-1, dtype=paddle.float32).astype(query.dtype)
         attn_weights = nn.functional.dropout(attn_weights, p=dropout, training=module.training)
 
-    attn_output = paddle.matmul(attn_weights, value)  # b h l l @ b h l d -> b h l d
-    attn_output = attn_output.transpose([0, 2, 1, 3])  # b h l d -> b l h d
+    attn_output = paddle.matmul(attn_weights, value)
+    attn_output = attn_output.transpose(1, 2).contiguous()
     attn_output = paddle.reshape(x=attn_output, shape=[0, 0, attn_output.shape[2] * attn_output.shape[3]])
 
     return attn_output, attn_weights
